Change stackPointer's type from Uint8 to Uint16
Because NES's stack pointer has its range of 0x100..0x1FF, though its capability is limited to 8-bit width.

+++ src/cpuRegisters.ts
@@ -5,7 +5,7 @@ export default class CpuRegisters {
   indexX: Uint8;
   indexY: Uint8;
   programCounter: Uint16;
-  stackPointer: Uint8;
+  stackPointer: Uint16;
   status: Uint8;
 
   constructor() {
